split a few more classes to their own files
