+++ TractorMod/TractorMod.csproj
@@ -35,7 +35,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Framework\PhthaloBlue.cs" />
+    <Compile Include="Framework\Save.cs" />
     <Compile Include="Framework\SaveCollection.cs" />
+    <Compile Include="Framework\ToolConfig.cs" />
     <Compile Include="Framework\Tractor.cs" />
     <Compile Include="Framework\TractorConfig.cs" />
     <Compile Include="Framework\TractorHouse.cs" />
